misc/HtmlTestSuite/results/test-0893.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>WebSockets: response header and close frame in same packet</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="../constants.sub.js"/>
<meta content="" name="variant"/>
<meta content="?wss" name="variant"/>
</head><body><div id="log"/>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/simple_handshake');
  ws.onmessage = t.unreached_func();
  ws.onopen = t.step_func(function(e) {
    ws.onclose = t.step_func(function(e) {
      assert_equals(e.wasClean, true);
      assert_equals(e.code, 1001);
      assert_equals(e.reason, 'PASS');
      ws.onclose = t.unreached_func('onclose should not be called twice');
      t.step_timeout(() =&gt; t.done(), 50);
    })
    ws.close();
  })
  ws.onclose = t.unreached_func('onclose should not be called before onopen');
});
</script>
</body></html>
